fix: remove CSRProtect
I understand that using the CSRFProtect may break some functionalities,
so for the time being I will remove it in order to continue the development.
However, I will make sure to fix this issue and re-implement the CSRFProtect
at a later stage. This will ensure that all functionalities are working properly
and that the application is secure.